requirements.txt
dash==2.17.0
dash-bootstrap-components==1.6.0
sentence-transformers>=2.6.0
numpy>=1.23
huggingface-hub>=0.23.0
pyspellchecker
